Add files via upload
Loop asking for input X and Y, returning Z and asks the user wants to continue or not. The loop invariant which expresses the relationship
between Aux, Z, and X, and gives a bound on Aux. The program passes the data and information flow analysis of Spark Ada and meaningful verification conditions are present.

This was developed by me as a solution to a High Integrity task from Swansea University.